requirements.txt
pillow
gooey